requirements.txt
Flask
Flask_Migrate
flask_sqlalchemy
mysql_connector_repackaged
pandas
pandas_stubs
Werkzeug
PyJWT
scikit-learn